Extract LazyColumn helpers
