+++ redwood-treehouse-lazylayout-schema/compose/build.gradle
@@ -1,0 +1,21 @@
+apply plugin: 'org.jetbrains.kotlin.multiplatform'
+apply plugin: 'app.cash.redwood.generator.compose'
+
+kotlin {
+  js {
+    browser()
+  }
+
+  sourceSets {
+    commonMain {
+      dependencies {
+        api projects.redwoodTreehouseLazylayoutSchema.widget
+      }
+    }
+  }
+}
+
+redwoodSchema {
+  source = projects.redwoodTreehouseLazylayoutSchema
+  type = 'app.cash.redwood.treehouse.lazylayout.RedwoodTreehouseLazyLayout'
+}
