+++ samples/emoji-search/schema/compose/build.gradle
@@ -12,6 +12,7 @@ kotlin {
     commonMain {
       dependencies {
         api(projects.samples.emojiSearch.schema.widget)
+        api(projects.redwoodTreehouseLazylayoutSchema.compose)
       }
     }
   }
