+++ settings.gradle
@@ -22,6 +22,7 @@ include ':redwood-treehouse-composeui'
 include ':redwood-treehouse-lazylayout-api'
 include ':redwood-treehouse-lazylayout-composeui'
 include ':redwood-treehouse-lazylayout-schema'
+include ':redwood-treehouse-lazylayout-schema:compose'
 include ':redwood-treehouse-lazylayout-schema:widget'
 include ':redwood-treehouse-lazylayout-uiview'
 include ':redwood-treehouse-lazylayout-view'
